+++ src/components/Main/Modal/SettingModal/BrowserTab/Account.vue
@@ -1,0 +1,45 @@
+<template>
+  <div>
+    <h3>アカウント</h3>
+    <form-button label="ログアウト" on-secondary @click="onLogoutClick" />
+    <form-button
+      label="全セッション破棄"
+      on-secondary
+      @click="onSessionDelete"
+    />
+  </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from '@vue/composition-api'
+import apis from '@/lib/api'
+import FormButton from '@/components/UI/FormButton.vue'
+
+export default defineComponent({
+  name: 'Account',
+  setup(props, context) {
+    const onLogoutClick = async () => {
+      await apis.logout()
+      context.root.$router.push('/')
+    }
+
+    const onSessionDelete = async () => {
+      // TODO: セッション表示と特定のセッション破棄とかする？
+      if (
+        window.confirm(
+          'ログイン中のセッションを全て破棄します。（実行するとログアウトされます）'
+        )
+      ) {
+        await apis.logout(undefined, true)
+        context.root.$router.push('/')
+      }
+    }
+    return { onLogoutClick, onSessionDelete }
+  },
+  components: {
+    FormButton
+  }
+})
+</script>
+
+<style lang="scss" module></style>
